+++ packages/patrol_cli/lib/src/commands/develop_service.dart
@@ -427,6 +427,7 @@ class DevelopService {
           clearTestSteps: clearTestSteps,
           stdin: _stdin,
           onLogEntry: onLogEntry,
+          openDevtools: openDevtools,
         );
     }